% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{build_usbe_url}
\alias{build_usbe_url}
\title{Build USBE enrollment data URL}
\usage{
build_usbe_url(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Character string with full URL
}
\description{
Constructs the URL for downloading enrollment data from USBE.
}
\keyword{internal}
