% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_usbe_enrollment}
\alias{download_usbe_enrollment}
\title{Download USBE Fall Enrollment data}
\usage{
download_usbe_enrollment(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
List with state, district (LEA), and school data frames
}
\description{
Downloads the Fall Enrollment by Grade Level and Demographics Excel file
from the USBE Data and Statistics portal and reads all relevant sheets.
}
\keyword{internal}
