% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{standardize_columns}
\alias{standardize_columns}
\title{Standardize column names from USBE Excel files}
\usage{
standardize_columns(df)
}
\arguments{
\item{df}{Data frame with raw column names}
}
\value{
Data frame with standardized column names
}
\description{
Converts raw USBE column names to standardized format.
}
\keyword{internal}
